dev-requirements.txt
Django==3.2
environs==9.5.0
Pillow==10.0.1
django-admin-sortable2==1.0.4
django-tinymce==3.6.1
requests==2.31.0
django-debug-toolbar==3.3.0
